Fixes int conversion warning
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinxbase@12658 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ src/libsphinxbase/util/pio.c
@@ -73,7 +73,7 @@ enum {
 static void
 guess_comptype(char const *file, int32 *ispipe, int32 *isgz)
 {
-    int k;
+    size_t k;
 
     k = strlen(file);
     *ispipe = 0;
@@ -206,7 +206,7 @@ fopen_compchk(const char *file, int32 * ispipe)
         return fh;
     else {
         char *tmpfile;
-        int k;
+        size_t k;
 
         /* File doesn't exist; try other compressed/uncompressed form, as appropriate */
         guess_comptype(file, ispipe, &isgz);
@@ -400,8 +400,8 @@ int32
 fread_retry(void *pointer, int32 size, int32 num_items, FILE * stream)
 {
     char *data;
-    uint32 n_items_read;
-    uint32 n_items_rem;
+    size_t n_items_read;
+    size_t n_items_rem;
     uint32 n_retry_rem;
     int32 loc;
